+++ app/download-bluesky-video/page.tsx
@@ -110,18 +110,29 @@ export default async function Page({ searchParams }: Props) {
             'text': item.a
           }
         }))
+      },
+      {
+        '@type': 'BreadcrumbList',
+        'itemListElement': [
+          {
+            '@type': 'ListItem',
+            'position': 1,
+            'name': menu.downloads,
+            'item': 'https://clipkeep.net/'
+          },
+          {
+            '@type': 'ListItem',
+            'position': 2,
+            'name': t.title,
+            'item': `https://clipkeep.net/download-bluesky-video?locale=${locale}`
+          }
+        ]
       }
     ]
   };
 
   return (
     <>
-      <BreadcrumbSchema 
-        items={[
-          { name: menu.downloads, item: '/' },
-          { name: t.title, item: `/download-bluesky-video?locale=${locale}` }
-        ]}
-      />
       <script
         type="application/ld+json"
         dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}